Implemented withdraw and deposit methods using Transaction.

+++ src/main/java/account/Transaction.java
@@ -2,10 +2,10 @@ package account;
 
 public class Transaction {
     private final String id;
-    private final int amount;
+    private final double amount;
     private final TransactionType type;
 
-    public Transaction(String id, int amount, TransactionType type) {
+    public Transaction(String id, double amount, TransactionType type) {
         this.id = id;
         this.amount = amount;
         this.type = type;
